+++ Ctrl 2K Engine 0.0.1.sln
@@ -37,9 +37,9 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AssetsCore", "Ctrl 2K Engin
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CtrlCore", "CtrlCore", "{EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CtrlMiscUtils", "Core\CtrlMiscUtils.vcxproj", "{385BC729-58FA-49D3-8609-0A33B280E111}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Petra", "Core\Petra\Petra.vcxproj", "{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CtrlMiscUtils", "Core\CtrlMiscUtils\CtrlMiscUtils.vcxproj", "{385BC729-58FA-49D3-8609-0A33B280E111}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -121,14 +121,6 @@ Global
 		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Release|x64.Build.0 = Release|x64
 		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Release|x86.ActiveCfg = Release|Win32
 		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Release|x86.Build.0 = Release|Win32
-		{385BC729-58FA-49D3-8609-0A33B280E111}.Debug|x64.ActiveCfg = Debug|x64
-		{385BC729-58FA-49D3-8609-0A33B280E111}.Debug|x64.Build.0 = Debug|x64
-		{385BC729-58FA-49D3-8609-0A33B280E111}.Debug|x86.ActiveCfg = Debug|Win32
-		{385BC729-58FA-49D3-8609-0A33B280E111}.Debug|x86.Build.0 = Debug|Win32
-		{385BC729-58FA-49D3-8609-0A33B280E111}.Release|x64.ActiveCfg = Release|x64
-		{385BC729-58FA-49D3-8609-0A33B280E111}.Release|x64.Build.0 = Release|x64
-		{385BC729-58FA-49D3-8609-0A33B280E111}.Release|x86.ActiveCfg = Release|Win32
-		{385BC729-58FA-49D3-8609-0A33B280E111}.Release|x86.Build.0 = Release|Win32
 		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Debug|x64.ActiveCfg = Debug|x64
 		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Debug|x64.Build.0 = Debug|x64
 		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Debug|x86.ActiveCfg = Debug|Win32
@@ -137,6 +129,14 @@ Global
 		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Release|x64.Build.0 = Release|x64
 		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Release|x86.ActiveCfg = Release|Win32
 		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Release|x86.Build.0 = Release|Win32
+		{385BC729-58FA-49D3-8609-0A33B280E111}.Debug|x64.ActiveCfg = Debug|x64
+		{385BC729-58FA-49D3-8609-0A33B280E111}.Debug|x64.Build.0 = Debug|x64
+		{385BC729-58FA-49D3-8609-0A33B280E111}.Debug|x86.ActiveCfg = Debug|Win32
+		{385BC729-58FA-49D3-8609-0A33B280E111}.Debug|x86.Build.0 = Debug|Win32
+		{385BC729-58FA-49D3-8609-0A33B280E111}.Release|x64.ActiveCfg = Release|x64
+		{385BC729-58FA-49D3-8609-0A33B280E111}.Release|x64.Build.0 = Release|x64
+		{385BC729-58FA-49D3-8609-0A33B280E111}.Release|x86.ActiveCfg = Release|Win32
+		{385BC729-58FA-49D3-8609-0A33B280E111}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -146,8 +146,8 @@ Global
 		{EB11A967-DAB8-4DE2-894C-F5014E62137E} = {A1C2A35E-4285-470E-AD90-009320CE8212}
 		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1} = {A1C2A35E-4285-470E-AD90-009320CE8212}
 		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6} = {26927F71-5167-41ED-8F46-A20AA59F0B40}
-		{385BC729-58FA-49D3-8609-0A33B280E111} = {EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}
 		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF} = {EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}
+		{385BC729-58FA-49D3-8609-0A33B280E111} = {EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {70BAD174-8BA7-4F92-9815-19B71A819840}
